+++ docs/changelog.rst
@@ -3,14 +3,20 @@
 Changelog
 =========
 
-Version 0.4.0 - Unreleased
+Version 0.4.0 - 2016/10/23
 --------------------------
 
-* Added a configuration registry and use it instead of directly getting config from ``settings.RSTVIEW_PARSER_FILTER_SETTINGS``;
+* Added a configuration registry and use it instead of directly getting/setting configurations from ``settings.RSTVIEW_PARSER_FILTER_SETTINGS``;
 * Now configurations can be added/updated/removed per application through registry interface;
 * Now ``settings.RSTVIEW_PARSER_FILTER_SETTINGS`` is only used as a startup configuration sets that are not needed to be tampered anymore;
 * Added discovering methods to discover configuration files from enabled applications;
-* Added tests for registry;
+* Added tests for registry and discovering;
+* Updated documentation to include configuration registry and all;
+* Removed some unused settings:
+
+  * ``RSTVIEW_PARSER_ENABLE_FILE_INSERTION``;
+  * ``RSTVIEW_PARSER_ENABLE_RAW_INSERTION``;
+  * ``RSTVIEW_PARSER_LANGUAGE_CODE``.
 
 Version 0.3.2 - 2016/07/26
 --------------------------